requirements.txt
asgiref==3.8.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
idna==3.6
psycopg2==2.9.9
PyJWT==2.9.0
requests==2.31.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.1.0
